# prioritize non node_modules