if we check non node_modules files first, we still get a significant speed up:

before: `12.93s user 0.68s system 137% cpu 9.862 total`
after: `12.22s user 0.67s system 139% cpu 9.248 total`