@bywo 2019-01-24
haha fair. it's a 0.7s difference. If you're running from the command line compiling one type at a time, it's not going to be huge. But my use-case is compiling programmatically (not via CLI) on a watch loop (i.e. recompile on file change). Those 0.7s differences will add up for me.