@bywo 2019-01-24
Maybe we can discuss your use case and figure out how to speed things up for you too!