@domoritz 2019-01-24
My use-case is https://github.com/vega/vega-lite/blob/master/package.json#L44. If you have suggestions, you could make us very happy.