@bywo 2019-01-24
whoops, just realized you had `--no-type-check` intentionally. let me remove my change there. I thought it was a typo