@bywo 2019-01-25
thanks for merging! do you mind publishing a new version?