@domoritz 2019-01-25
Done (release will be available in a few minutes).